+++ lib/rmodbus/errors.rb
@@ -1,6 +1,6 @@
 # RModBus - free implementation of ModBus protocol on Ruby.
 #
-# Copyright (C) 2008  Timin Aleksey
+# Copyright (C) 2008-2011  Timin Aleksey
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -14,7 +14,6 @@
 module ModBus
 
   module Errors
-
     class ProxyException < StandardError
     end
 
@@ -45,6 +44,13 @@ module ModBus
     class ModBusTimeout < ModBusException
     end
 
+    class ResponseMismatch < ModBusException
+      attr_reader :request, :response
+      def initialize(msg, request, response)
+        super(msg)
+        @request = request
+        @response = response
+      end 
+    end
   end
-
 end