+++ lib/rmodbus/tcp_server.rb
@@ -1,6 +1,6 @@
 # RModBus - free implementation of ModBus protocol on Ruby.
 #
-# Copyright (C) 2008  Timin Aleksey
+# Copyright (C) 2008-2011  Timin Aleksey
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -24,7 +24,7 @@ module ModBus
   #   srv.debug = true
   #   srv.start
 	class TCPServer < GServer
-		include Common
+		include Debug
 		include Server 
     
     # Init server
@@ -43,23 +43,21 @@ module ModBus
     # Serve requests
     # @param [TCPSocket] io socket
 		def serve(io)
-			loop do
+			while not stopped? 
 				req = io.read(7)
-				if req[2,2] != "\x00\x00" or req.getbyte(6) != @uid
-					io.close
-					break
-				end
+				if req[2,2] == "\x00\x00" or req.getbyte(6) == @uid
+          tr = req[0,2]
+          len = req[4,2].unpack('n')[0]
+          req = io.read(len - 1)
+          log "Server RX (#{req.size} bytes): #{logging_bytes(req)}"
 
-				tr = req[0,2]
-				len = req[4,2].unpack('n')[0]
-				req = io.read(len - 1)
-				log "Server RX (#{req.size} bytes): #{logging_bytes(req)}"
+          pdu = exec_req(req, @coils, @discrete_inputs, @holding_registers, @input_registers)
 
-				pdu = exec_req(req, @coils, @discrete_inputs, @holding_registers, @input_registers)
-
-				resp = tr + "\0\0" + (pdu.size + 1).to_word + @uid.chr + pdu
-				log "Server TX (#{resp.size} bytes): #{logging_bytes(resp)}"
-				io.write resp
+          resp = tr + "\0\0" + (pdu.size + 1).to_word + @uid.chr + pdu
+          log "Server TX (#{resp.size} bytes): #{logging_bytes(resp)}"
+          io.write resp
+        end
+        sleep(0.01)
 			end
 		end
 	end