+++ spec/rtu_server_spec.rb
@@ -1,13 +1,12 @@
 require 'rmodbus'
-include ModBus
 
-describe RTUServer do
+describe ModBus::RTUServer do
   before do
     @sp = mock "SerialPort"
     SerialPort.should_receive(:new).with('/dev/ttyS0', 4800, 7, 2, SerialPort::NONE).and_return(@sp)
     @sp.stub!(:read_timeout=)
 
-    @server = RTUServer.new('/dev/ttyS0', 4800, 1, :data_bits => 7, :stop_bits => 2)
+    @server = ModBus::RTUServer.new('/dev/ttyS0', 4800, 1, :data_bits => 7, :stop_bits => 2)
     @server.coils = [1,0,1,1]
     @server.discrete_inputs = [1,1,0,0]
     @server.holding_registers = [1,2,3,4]