+++ NEWS.md
@@ -1,3 +1,20 @@
+Next Release 1.1.0
+===================================
+1. Fixed issue [#12](https://github.com/flipback/rmodbus/issues/12). Added option Slave#raise_exception_on_mismatch to turn to check response and raise exception
+   if it's mismatch.
+2. Added pass options :debug, :raise_exception_on_mismatch, :read_retry_timeout, :read_retries from clients to slaves
+  `ruby
+    @cl.debug = true
+    @cl.with_slave(1) do |slave_1|
+      slave_1.debug #=> true
+    end
+    @cl.with_slave(2) do |slave_2|
+      slave_2.debug = false
+      slave_2.debug #=> false
+    end
+  ` 
+3. Deleted dependency with `serialport` gem. Install it manual for using RTU
+
 2011-08-10 Release 1.0.4
 ====================================
 1. Fixed issue [#11](https://github.com/flipback/rmodbus/issues/11)
@@ -27,7 +44,7 @@ New API for client part of library
 ---------------------------------------
 
 Example:
-
+  `ruby
     require 'rmodbus'
 
     ModBus::TCPClient.new('127.0.0.1', 8502) do |cl|
@@ -45,7 +62,7 @@ Example:
         slave.holding_registers[16..20] = [1, 2, 3, 4, 5]
       end
     end
-
+  `
 for more information [see](http://rdoc.info/gems/rmodbus/1.0.0/frames)
 
 Conversion to/from 32bit registers
@@ -54,7 +71,7 @@ Conversion to/from 32bit registers
 Some modbus devices use two registers to store 32bit values.
 RModbus provides some helper functions to go back and forth between these two things when reading/writing.
 The built-in examples assume registers in a particular order but it's trivial to change.
-
+  `ruby
     # Reading values in multiple registers (you can read more than 2 and convert them all so long as they are in multiples of 2)
     res = slave.holding_registers[0..1]
     res.inspect => [20342, 17344]
@@ -66,7 +83,7 @@ The built-in examples assume registers in a particular order but it's trivial to
     cl.holding_registers[0..1] => [20342, 17344]
     cl.holding_registers[2..3] = [384.620788574219].from_32f
     cl.holding_registers[2..3] => [20342, 17344]
-
+  `
 Support JRuby
 --------------------------------------
 Now you could use RModBus on JRuby without RTU implementation.